+++ src/manifest.ts
@@ -8,6 +8,7 @@ const manifest: ManifestType = {
   description: packageJson.description,
   author: packageJson.author,
   options_page: "src/pages/options/index.html",
+  background: { service_worker: "src/pages/background/index.js" },
   action: {
     default_title: "Peek History",
     default_popup: "src/pages/popup/index.html",
@@ -24,7 +25,14 @@ const manifest: ManifestType = {
     "32": "icon32.png",
     "16": "icon16.png",
   },
-  permissions: ["fontSettings", "tabs", "history", "storage", "sessions"],
+  permissions: [
+    "fontSettings",
+    "tabs",
+    "history",
+    "storage",
+    "sessions",
+    "favicon",
+  ],
   web_accessible_resources: [
     {
       resources: ["icon128.png"],
@@ -36,7 +44,7 @@ const manifest: ManifestType = {
   },
   incognito: "not_allowed",
   commands: {
-    _execute_browser_action: {
+    _execute_action: {
       suggested_key: {
         default: "Ctrl+Shift+H",
         mac: "MacCtrl+Shift+H",