+++ src/pages/popup/redux/history/history-slice.ts
@@ -20,11 +20,13 @@ export interface IHistoryItem {
 interface historyState {
   loading: boolean;
   items: IHistoryItem[];
+  remaining: boolean;
 }
 
 const initialState: historyState = {
   loading: true,
   items: [],
+  remaining: true,
 };
 
 const historySlice = createSlice({
@@ -39,10 +41,17 @@ const historySlice = createSlice({
         state.loading = false;
       }
     ),
+      // Update/Append history
       builder.addCase(
         addHistory.fulfilled,
-        (state, { payload }: PayloadAction<IHistoryItem[]>) => {
-          state.items.push(...payload);
+        (
+          state,
+          {
+            payload,
+          }: PayloadAction<{ history: IHistoryItem[]; remaining: boolean }>
+        ) => {
+          state.remaining = payload.remaining;
+          state.items.push(...payload.history);
         }
       ),
       builder.addCase(
@@ -62,9 +71,12 @@ const historySlice = createSlice({
         state.loading = false;
         state.items = [];
       }),
-      builder.addCase(deleteRange.fulfilled, (state) => {
-        state.items = [];
-      });
+      builder.addCase(
+        deleteRange.fulfilled,
+        (state, { payload }: PayloadAction<IHistoryItem[]>) => {
+          state.items = payload;
+        }
+      );
   },
 });