+++ src/pages/popup/redux/ui/thunks.ts
@@ -1,0 +1,10 @@
+import { createAsyncThunk } from "@reduxjs/toolkit";
+
+export const getShowUpdatePopup = createAsyncThunk(
+  "ui/getShowUpdatePopup",
+  async () => {
+    const data = await chrome.storage.sync.get("showUpdatePopup");
+    const showUpdatePopup = data?.showUpdatePopup;
+    return showUpdatePopup;
+  }
+);