+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ## Changelog
 
-### [2.0.0] - DD-MM-2022
+### [2.0.0] - 01-03-2023
 - Switched to Redux for state management
 - Switched to Vite for bundling
 - Upgraded to React 18
+- Increased favicon quality with New Favicon API
+- Added option to delete history in time ranges
+- Improved keyboard navigation and shortcuts
+- Changed shortcuts modifier from CTRL to Alt/Option
+- Added setting to set tab (current or new blank) for opening URLs
+- Added hide time and accent font color settings in options
+- Improved UI and UX
 
 ### [1.0.8] - 08-04-2021
 - Switch from node-sass to sass (dart sass)